fix front page tests
