+++ src/app/pages/FrontPage/index.test.jsx
@@ -8,7 +8,7 @@ import { ToggleContextProvider } from '#contexts/ToggleContext';
 import frontPageDataPidgin from '#data/pidgin/frontpage/index-light';
 import pidginMostReadData from '#data/pidgin/mostRead';
 import getInitialData from '#app/routes/home/getInitialData';
-import { FrontPage } from '..';
+import FrontPage from '.';
 
 const requestContextData = {
   pageType: 'frontPage',
